src/chatbot/RecruiterAssistantChatbot.java
package chatbot;

import java.util.Scanner;

public class RecruiterAssistantChatbot {

    public static void main(String[] args) {
        Scanner scanner = new Scanner(System.in);
        int choice;

        do {
            System.out.println("Select an option:");
            System.out.println("1. Assist job seeker in finding job vacancies");
            // ... (add other options)

            
            System.out.println("0. Exit");

            choice = scanner.nextInt();
            scanner.nextLine(); // Consume the newline character

            switch (choice) {
                case 1:
                    assistFindingJobVacancies();
                    break;
                // ... (add cases for other options)

               
                case 0:
                    System.out.println("Exiting Recruiter Assistant Chatbot. Goodbye!");
                    break;
                default:
                    System.out.println("Invalid choice. Please try again.");
            }
        } while (choice != 0);

        scanner.close();
    }

    private static void assistFindingJobVacancies() {
        System.out.println("Enter your skills:");
        try (Scanner scanner = new Scanner(System.in)) {
			String skills = scanner.nextLine();
			// Assume a JobVacancyService class that fetches job vacancies based on skills.
			JobVacancyService jobVacancyService = new JobVacancyService();
			jobVacancyService.findJobVacancies(skills);
		}
    }

    // ... (implement other methods for each user story)

  
}

class JobVacancyService {
    // Assume this class interacts with a job database or API.
    public void findJobVacancies(String skills) {
        // Implement logic to fetch job vacancies based on skills.
        System.out.println("Fetching job vacancies for skills: " + skills);
        // Display the job vacancies to the user.
    }
}
